+++ example/android/app/src/main/kotlin/com/security/flutter_app_guard/app_guard_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.security.app_guard.app_guard_example
+package com.security.flutter_app_guard.app_guard_example
 
 import io.flutter.embedding.android.FlutterActivity